+++ packages/app/components/EmptyPage/NoNotification.tsx
@@ -1,7 +1,7 @@
 import { View, Text, VStack } from "native-base";
 import React from "react";
 import LottieView from "lottie-react-native";
-import Animated, { FadeInDown, FadeIn } from "react-native-reanimated";
+import Animated, { FadeInDown, FadeIn, FadeOut } from "react-native-reanimated";
 
 type Props = {};
 
@@ -11,7 +11,8 @@ export const NoNotification = (props: Props) => {
   return (
     <VStack flex={1} justifyContent={"center"} alignItems={"center"}>
       <AnimatedLottieView
-        entering={FadeIn.delay(80)}
+        entering={FadeIn.delay(100)}
+        exiting={FadeOut}
         // ref={animation => {
         //   this.animation = animation;
         // }}
@@ -28,7 +29,10 @@ export const NoNotification = (props: Props) => {
         style={{ marginBottom: 60 }}
       />
       {/* REVIEW: RNR LAYOUT ANIMATION */}
-      <Animated.View entering={FadeInDown.delay(100).springify()}>
+      <Animated.View
+        entering={FadeInDown.delay(100).springify()}
+        exiting={FadeOut}
+      >
         <Text pt={80} fontSize={18} fontWeight={"600"} color={"blueGray.500"}>
           No Notifications
         </Text>