+++ packages/app/components/EmptyPage/index.ts
@@ -4,3 +4,4 @@ export * from "./Empty";
 export * from "./Loading";
 export * from "./Download";
 export * from "./DayAndNight";
+export * from "./TryAgain";